Feature/KoinImplementation: - Refactoring - Add errors to HomeRemoteDataSource response
