+++ .idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-06-04T18:58:57.160033Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-09-04T15:18:58.641507Z" />
   </component>
 </project>
