+++ business/src/main/java/com/compose/business/common/model/AppError.kt
@@ -1,5 +1,5 @@
 package com.compose.business.common.model
 
 enum class AppError {
-    UNKNOWN, NETWORK_ERROR
+    UNKNOWN, NETWORK_ERROR, EMPTY_RESPONSE
 }
