+++ platform/src/main/java/com/compose/platform/MainActivity.kt
@@ -3,6 +3,7 @@ package com.compose.platform
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import com.compose.platform.navigation.StartComposableApp
 import com.compose.platform.theme.CryptoTheme
 
 class MainActivity : ComponentActivity() {
